+++ client/tws_foundation_client/lib/src/sets/truck_external.dart
@@ -79,18 +79,20 @@ final class TruckExternal implements CSMSetInterface {
     if(common < 0) results.add(CSMSetValidationResult(kCommon, 'Common pointer must be equal or greater than 0', 'pointerHandler()'));
     if(status < 0) results.add(CSMSetValidationResult(kStatus, 'Status pointer must be equal or greater than 0', 'pointerHandler()'));
     
-    if(carrier.isEmpty || carrier.length > 100) results.add(CSMSetValidationResult(kCarrier, "Carrier length must be between 1 and 100", "strictLength(1, 100)"));
+    if(carrier.trim().isEmpty || carrier.length > 100) results.add(CSMSetValidationResult(kCarrier, "Carrier length must be between 1 and 100", "strictLength(1, 100)"));
 
     if(mxPlate != null){
-      if((mxPlate!.length < 5 || mxPlate!.length > 12) && mxPlate!.isNotEmpty){
-        results.add(CSMSetValidationResult(kMxPlate, "External Truck Mexican plate length must be between 8 and 12", "strictLength(8, 12)"));
+      if(mxPlate!.trim().isEmpty) results.add(CSMSetValidationResult(kMxPlate, "Elimine los espacios en blanco de las placas mexicanas en el trailer externo.", "emptyString()"));
+      if((mxPlate!.length < 5 || mxPlate!.length > 12) && mxPlate!.trim().isNotEmpty){
+        results.add(CSMSetValidationResult(kMxPlate, "Las placas mexicanas del camión externo deben tener una longitud de entre 5 y 12 caracteres.", "strictLength(5, 12)"));
       }else{
         isPlate = true;
       }
     }
     if(usaPlate != null){
-      if((usaPlate!.length < 5 || usaPlate!.length > 12) && usaPlate!.isNotEmpty){
-        results.add(CSMSetValidationResult(kUsaPlate, "External Truck American plate length must be between 8 and 12", "strictLength(8, 12)"));
+      if(usaPlate!.trim().isEmpty) results.add(CSMSetValidationResult(kUsaPlate, "Elimine los espacios en blanco de las placas americanas en el trailer externo.", "emptyString()"));
+      if((usaPlate!.length < 5 || usaPlate!.length > 12) && usaPlate!.trim().isNotEmpty){
+        results.add(CSMSetValidationResult(kUsaPlate, "Las placas americanas del camión externo deben tener una longitud de entre 5 y 12 caracteres.", "strictLength(5, 12)"));
       }else{
         isPlate = true;
       }
@@ -119,20 +121,22 @@ final class TruckExternal implements CSMSetInterface {
   }){
     String? uPlate = usaPlate ?? this.usaPlate;
     if(usaPlate != null){
-      if(usaPlate.isEmpty){
+      if(usaPlate.trim().isEmpty){
         uPlate = null;
       }
     }
     String? mPlate = mxPlate ?? this.mxPlate;
     if(mxPlate != null){
-      if(mxPlate.isEmpty){
+      if(mxPlate.trim().isEmpty){
         mPlate = null;
       }
     }
    
     String? v = vin ?? this.vin;
-    if(v == "") v = null;
-    
+    if(v != null){
+      if(v.trim().isEmpty) v = null;
+    }
+
     return TruckExternal(id ?? this.id, status ?? this.status, common ?? this.common, vin, carrier ?? this.carrier, mPlate, uPlate, truckCommonNavigation ?? this.truckCommonNavigation, statusNavigation ?? this.statusNavigation);
   }
 }