+++ client/tws_foundation_client/lib/src/sets/yard_log.dart
@@ -18,6 +18,7 @@ final class YardLog implements CSMSetInterface {
   static const String kGName = "gname";
   static const String kFromTo = "fromTo";
   static const String kSeal = "seal";
+  static const String kSealAlt = "sealAlt";
   static const String kDamage = "damage";
   static const String kTtPicture = "ttPicture";
   static const String kDmgEvidence = "dmgEvidence";
@@ -49,6 +50,7 @@ final class YardLog implements CSMSetInterface {
   String gName = "";
   String fromTo = "";
   String? seal;
+  String? sealAlt;
   bool damage = false;
   String ttPicture = "";
   String? dmgEvidence;
@@ -63,7 +65,7 @@ final class YardLog implements CSMSetInterface {
   Account? accountNavigation; //Only for local handle.
 
   YardLog(this.id, this.entry, this.truck, this.truckExternal, this.trailer, this.trailerExternal, this.loadType, this.section, this.driver, this.driverExternal,
-  this.guard, this.gName, this.fromTo, this.seal, this.damage, this.ttPicture, this.dmgEvidence, this.driverNavigation, this.driverExternalNavigation, this.truckNavigation, this.truckExternalNavigation,
+  this.guard, this.gName, this.fromTo, this.seal, this.sealAlt, this.damage, this.ttPicture, this.dmgEvidence, this.driverNavigation, this.driverExternalNavigation, this.truckNavigation, this.truckExternalNavigation,
   this.trailerNavigation, this.trailerExternalNavigation, this.loadTypeNavigation, this.sectionNavigation, this.accountNavigation, { 
     DateTime? timestamp,
   }){
@@ -85,6 +87,7 @@ final class YardLog implements CSMSetInterface {
     String gName = json.get('gName');
     String fromTo = json.get('fromTo');
     String? seal = json.getDefault('seal', null);
+    String? sealAlt = json.getDefault('sealAlt', null);
     bool damage = json.get('damage');
     String ttPicture = json.get('ttPicture');
     String? dmgEvidence = json.getDefault('dmgEvidence', null);
@@ -130,7 +133,7 @@ final class YardLog implements CSMSetInterface {
       sectionNavigation = deserealize<Section>(rawNavigation, decode: SectionDecoder());
     }
 
-    return YardLog(id, entry, truck, truckExternal, trailer, trailerExternal, loadType, section, driver, driverExternal, guard, gName, fromTo, seal, damage,
+    return YardLog(id, entry, truck, truckExternal, trailer, trailerExternal, loadType, section, driver, driverExternal, guard, gName, fromTo, seal, sealAlt, damage,
     ttPicture, dmgEvidence, driverNavigation, driverExternalNavigation, truckNavigation, truckExternalNavigation, trailerNavigation, trailerExternalNavigation, loadTypeNavigation, sectionNavigation, null, timestamp: timestamp, );
   }
 
@@ -153,6 +156,7 @@ final class YardLog implements CSMSetInterface {
       kGName: gName,
       kFromTo: fromTo,
       kSeal: seal,
+      kSealAlt: sealAlt,
       kDamage: damage,
       kTtPicture: ttPicture,
       kDmgEvidence: dmgEvidence,
@@ -172,13 +176,17 @@ final class YardLog implements CSMSetInterface {
   List<CSMSetValidationResult> evaluate() {
     List<CSMSetValidationResult> results = <CSMSetValidationResult>[];
     if(ttPicture.isEmpty ) results.add(CSMSetValidationResult(kTtPicture, "Debe tomar una foto del camión con el remolque.", "strictLength(1, max)"));
-    if(gName.isEmpty || gName.length > 100) results.add(CSMSetValidationResult(kName, "El nombre del guardia no debe exeder los 100 caracteres y no debe estar vacio.", "strictLength(1,100)"));
-    if(fromTo.isEmpty || fromTo.length > 100) results.add(CSMSetValidationResult(kFromTo, "Debe indicar de donde viene (o a donde va el camión). Maximo 25 caracteres.", "strictLength(1,25)"));
+    if(gName.trim().isEmpty || gName.length > 100) results.add(CSMSetValidationResult(kName, "El nombre del guardia no debe exeder los 100 caracteres y no debe estar vacio.", "strictLength(1,100)"));
+    if(fromTo.trim().isEmpty || fromTo.length > 100) results.add(CSMSetValidationResult(kFromTo, "Debe indicar de donde viene (o a donde va el camión). Maximo 25 caracteres.", "strictLength(1,25)"));
     if(seal != null){
       if(trailerNavigation == null && trailerExternalNavigation == null) results.add(CSMSetValidationResult(kSeal, "Se ingreso un sello pero no un relmolque, seleccione alguno.", "FieldConflict()"));
-      if(seal!.isEmpty || seal!.length > 64) results.add(CSMSetValidationResult(kSeal, "El campo del sello no contiene un texto valido. Maximo 64 caracteres.", "strictLength(1,64)"));
+      if(seal!.trim().isEmpty || seal!.length > 64) results.add(CSMSetValidationResult(kSeal, "El campo del sello no contiene un texto valido. Maximo 64 caracteres.", "strictLength(1,64)"));
+    }
+    if(sealAlt != null){
+      if(sealAlt!.trim().isEmpty || sealAlt!.length > 64) results.add(CSMSetValidationResult(kSealAlt, "El campo del sello #2 (alternativo) es muy largo. Maximo 64 caracteres.", "strictLength(1,64)"));
     }
     if(seal == null && (trailerNavigation != null || trailerExternalNavigation != null)) results.add(CSMSetValidationResult(kSeal, "Debe agregar el campo de sello. De lo contrario seleccione el tipo de carga Botado.", "FieldConflict()"));
+    
     if(section < 0) results.add(CSMSetValidationResult(kSection, 'Debe seleccionar la seccion.', 'pointerHandler()'));
     if(loadType < 0) results.add(CSMSetValidationResult(kLoadType, 'Debe seleccionar el tipo de carga.', 'pointerHandler()'));
 
@@ -239,6 +247,7 @@ final class YardLog implements CSMSetInterface {
     String? gName,
     String? fromTo,
     String? seal,
+    String? sealAlt,
     bool? damage,
     String? ttPicture,
     String? dmgEvidence,
@@ -253,12 +262,23 @@ final class YardLog implements CSMSetInterface {
     Account? accountNavigation
   }){
     String? sea = seal ?? this.seal;
-    if(sea == "") sea = null;
+    if(sea != null){
+      if(sea.trim().isEmpty) sea = null;
+    }
+
+    String? sealt = sealAlt ?? this.sealAlt;
+    if(sealt != null){
+      if(sealt.trim().isEmpty) sealt = null;
+    }
 
     String? dmgEv = dmgEvidence ?? this.dmgEvidence;
-    if(dmgEv == "") dmgEv = null;
+    if(dmgEv != null){
+      if(dmgEv.trim().isEmpty) dmgEv = null;
+    }
 
-    if(gName == "") this.guard = 0;
+    if(gName != null){
+      if(gName.trim().isEmpty) this.guard = 0;
+    }
 
     LoadType? load = loadTypeNavigation ?? this.loadTypeNavigation;
     if(loadType == 0) load = null;
@@ -310,7 +330,7 @@ final class YardLog implements CSMSetInterface {
     
     return YardLog(id ?? this.id, entry ?? this.entry, truckIndex, truckExtIndex, trailerIndex, trailerExtIndex, 
     loadType ?? this.loadType, section ?? this.section, driverIndex,driverExtIndex, guard ?? this.guard, 
-    gName ?? this.gName, fromTo ?? this.fromTo, sea, damage ?? this.damage, ttPicture ?? this.ttPicture, dmgEv, driverNav, 
+    gName ?? this.gName, fromTo ?? this.fromTo, sea, sealt,damage ?? this.damage, ttPicture ?? this.ttPicture, dmgEv, driverNav, 
     driverExtNav, truckNav, truckExtNav, trailerNav, trailerExtNav, load, sect, accountNavigation ?? this.accountNavigation);
   }